Extract and export the `bytes` serializer
