+++ packages/umi-serializer-beet/src/index.ts
@@ -1,4 +1,5 @@
 export * from './createBeetSerializer';
+export * from './bytes';
 export * from './errors';
 export * from './plugin';
 export * from './pubkey';
