+++ packages/umi-serializer-data-view/src/index.ts
@@ -1,4 +1,5 @@
 export * from './createDataViewSerializer';
+export * from './bytes';
 export * from './errors';
 export * from './plugin';
 export * from './pubkey';
